move get_session to session.py
