+++ src/sympc/protocol/fss/fss.py
@@ -20,13 +20,13 @@ import torch as th
 import torchcsprng as csprng  # type: ignore
 
 from sympc.protocol.protocol import Protocol
+from sympc.session import get_session
 from sympc.store import CryptoPrimitiveProvider
 from sympc.store import register_primitive_generator
 from sympc.store import register_primitive_store_add
 from sympc.store import register_primitive_store_get
 from sympc.tensor import MPCTensor
 from sympc.tensor import ShareTensor
-from sympc.utils import get_session
 from sympc.utils import parallel_execution
 
 ttp_generator = csprng.create_random_device_generator()
