+++ src/sympc/session/__init__.py
@@ -1,6 +1,7 @@
 """Session class and utility functions used in conjunction with the session."""
 
 from .session import Session
+from .session import get_session
 from .session_manager import SessionManager
 
-__all__ = ["Session", "SessionManager"]
+__all__ = ["Session", "get_session", "SessionManager"]
